Added t3061 - another tricky circular group dependency.
This one is similar to t6045, where the circular dependency arises because of a normal link that was created during a previous update, which is no longer valid because a sticky link (on the group) is being removed. The group makes it trickier because we don't know necessarily which normal links are now invalid. We just remove all normal links on generated files and re-run the command instead.