Added t3064 - yet another non-circular dependency test.
This test is the reason for moving the circular dependency checking among groups until after all Tupfiles are parsed. Since there is no ordering between when new commands are added to groups and when old ones are removed, we can't know if the group graph is valid or not until after all Tupfiles have been parsed.