Fix t4082 - make sure updating mtime on the Tupfile doesn't cause a refactoring error.
I thought sqlite3_changes() was the changes for the transaction, but really it's just the changes introduced by the last update/insert/delete statement. During the scan stage, we are updating the mtime on the Tupfile node, which can add changes to the db. We really should be using sqlite3_total_changes(), and checking if that changes at all during the parse stage.