windows: Disable hot patching for WinXP
This should fix gcc -flto on WinXP with MinGW. The problem is that for some reason hot patching breaks the DeleteFile hooks, so the unlink() in gcc's lto-wrapper fails when it shouldn't. I'm not sure why the DeleteFile hooks aren't wrapped properly, but disabling hot patching fixes the issue.

Currently all hot patching seems to buy us is the fact that FindFirstFile can be hooked in Windows 7. The iat patching doesn't hook it properly, and trying to hook NtQueryDirectoryFile() ends up breaking things unexpectedly (even if the hook is just a pass-through).

So the simplest fix for now is to just disable hot-patching in WinXP - ideally there'd be a simpler hooking mechanism that would work across Windows versions and 32/64-bit, but that is a larger issue.