lua: Re-use execute_rule().
This means we are converting a lua table to a string, which tup then chunks up into a path_list/name_list. It's a little wonky, but the parser code isn't very well separated. Ideally we'd be creating (probably) a path_list, and passing that in instead. This at least shares most of the same code, so we don't have to do things twice.