Cache results of tup_db_get_links for stickies.
Since these can potentially be re-used many times, we can cache the result in the tup_entry. They are updated automatically as stickies are added/removed, except in the case where nodes are removed, since they remove all stickies for that node. In that case we simply invalidate the cache and reload from the db next time they are needed.