Added t4102 - delay gitignore processing to handle outputting in different dirs.
If we generate a gitignore file in one directory, and then later parse another directory that writes files to the first, they won't be included in the gitignore file. Instead we need to write the gitignore files after all directories have been parsed so we have the entire file manifest for the directory.