Added t4114 - add generated dirs to .gitignore files.
Another option would be to create a .gitignore file in the generated directory itself, but that is much trickier to manage because we don't have a Tupfile that has the .gitignore directive. It becomes difficult to know when the .gitignore file should be removed, since its existence would be dependent on having a .gitignore directive in any of the src directories. Instead we can just add the generated dir to the .gitignore file of its parent directory. Unfortunately this does mean you need to create a Tupfile in the parent dir.