Added t4117 - refactoring dependent dirs.
When we build the create graph, we load all directories in create_list, and those that are dependent on them. However, we call tup_db_unflag_create() on all nodes, even if they aren't in the create_list. SQLite wasn't counting these as modifications, but tup was adding them to the expected_changes counter, so the refactor check was failing.