Added t3072 - remove a group after it's been used in %r.
When using a group as an input with %r, we get a normal link from the group to the command. However, the group ghost check was only checking normal links to the group, so when the group was removed, the normal link would still be present and cause tup to report an error since the node no longer existed.