Added t4121, t4122, t4123 - convert normal->generated dir
If we remove a normal file, and all that's left are generated files from other directories, we should become a generated dir. Since this happens really in tup_del_id_type(), but we don't know the parsing results until after parsing happens, we need to essentially have a followup of every parsed Tupfile to check if it should become a generated dir.

We also need to have the conversion rule be: our dir has no normal files/directories, and we have either a generated subdir or a generated file from another directory.