Added t4131, t4132 - /proc tests in Linux.
If we don't bind-mount /proc into the fuse FS, then things like /proc/PID/exe return a .tup/mnt path (which is wrong), and /proc/meminfo doesn't contain any information, which some processes need.

This fixes issues with using an rpath that has $ORIGIN in it, since that relies on /proc/PID/exe to get the path, as well as running javac in a chroot environment, since it uses /proc/meminfo to determine some VM sizing.