Fix race in t7057.
We are using 'touch foo/out.txt' to get the monitor to chdir into foo/, but since that may be considered a file overwrite if timestamps are different, when the Tupfile now writes to bar/out.txt, it may not delete foo/out.txt since it has been converted to a normal file. Instead we can add the sleep so the file is always converted to a normal file and check that it still exists.