Added t4145 - make sure we can't use inputs from a directory we didn't write to.
A Tupfile can use inputs from a generated directory if it wrote to them, but not use inputs that another Tupfile wrote to that directory (unless it uses a group). If the Tupfile doesn't write to the directory at all then it may not exist at parse-time, so we need to get an error message.