Added t4147 - initial test for ^o flag to skip commands.
If the ^o flag is present, all outputs are moved to .tup/tmp before running instead of unlinking. After the command completes, we can check them to see if they are equal. If they are, we can skip future commands that depend on those files (unless some other input has changed).