Added t4150 - make sure we can hit the max fd limit still close/open.
This mimics what gold does - it will open a ton of files, hit the fd limit, and then try to close one before opening another. In the fuse fs, the flush() call was trying to dup another file descriptor, which was failing since we didn't have any left. This close(dup()) doesn't seem to do anything for us, so pull it out.