fuse: Use fuse_main() instead of fuse_mount/fuse_new to work with gold.
The gold linker is broken and picks the wrong version of the fuse_new() function, so the wrong argument parsing is used and results in the error:

fuse: unknown option `'
fuse_new: Operation not permitted

By calling fuse_main() directly we can avoid this issue. Of course this means we are now desynchronized wrt setting up the file-system and trying to use it. This was already a problem on OSX and FreeBSD anyway, so it will be better to have a single approach that works for all. I suspect the init() function usage will also only work on Linux, but we can add a sleep/try to access a dummy file in the file-system polling loop if necessary.