osx: Fix mount race condition and master_fork wait().
For some reason, the init() callback isn't sufficient for determining when the fs is fully ready to go, so the chdir() into the virtual fs fails unless we wait for it to be fully initialized. We do this by polling for the @tup@ virtual directory.

Also, OSXFUSE is setting signal(SIGCHLD, SIGIGN), which causes the waitpid() on the master_fork to fail. We set this back to SIGDFL so that works properly.