Update `test/tup.sh` to accommodate init, upd
None of the tests failed (on OS X at least), but these changes eliminate
the large numbers of warnings that printed.

Test t0001 is updated to create a Tupfile.ini in the custom init
directory that it creates.

Test t7052 is updated to expect the monitor to run twice, as it detects
the call to `touch Tupfile.ini` from the test framework.