Added t4152 - max files test with mmap.
In this case we open the files and then mmap() them. When we close the last file, we don't munmap() it, so FUSE still thinks it is open. In this case we still want to be able to open & read/write files. This is done by keeping track of how many files are open by the sub-process, and if it goes over a limit, we use the fusexmp model of doing an open()/read()/close() on a read call, rather than keeping the fd open in FUSE. This hurts performance in cases where many file descriptors are open, but at least keeps the system running (in particular for the gold linker).