Remove individual st->lock and use statuslock.
For some reason, when locking the finfo struct in server.c, helgrind will complain that the pthread_mutex_lock() on finfo->lock conflicts with pthread_mutex_unlock() in child_wait_notifer():

==19453== Possible data race during write of size 4 at 0xAF65A60 by thread #12
==19453== Locks held: none
==19453==    at 0x4C2FC73: pthread_mutex_lock (in /usr/lib/valgrind/vgpreload_helgrind-amd64-linux.so)
==19453==    by 0x41FEAC: finfo_lock (file.c:79)
...
==19453== This conflicts with a previous read of size 4 by thread #2
==19453== Locks held: 1, at address 0x75EB80
==19453==    at 0x4E4107C: __pthread_mutex_unlock_usercnt (pthread_mutex_unlock.c:52)
==19453==    by 0x4C30373: pthread_mutex_unlock (in /usr/lib/valgrind/vgpreload_helgrind-amd64-linux.so)

This is quite nonsensical, since they are two different locks. Using just the statuslock seems to avoid the issue.