Don't read all normal links while parsing.
Since we only need to check if a sticky link is also a normal link when the link is being removed, it's silly to always load all of the normal links during parse time. For directories that have commands with lots of inputs, doing so can really slow down parsing.