Fix Windows lock indication.
Apparently if the shared lock file was already locked, the open() from another tup process would fail with a "Bad pipe". Instead it seems we need to use the CreateFile() interface so that we can do the tup_try_flock() call. This required changed the tup_flock interface to use a tup_lock_t type so that it can be an int for POSIX and a HANDLE for Windows.

Also added flushing to the Windows printf wrapper so messages actually show up when they are printed.