Support for extra_outputs in Lua frule.
This is similar to the already supported extra_inputs. The files supplied in
extra_outputs are not substituted in %o. I also added a warning to the
documentation since the API for extra_input/extra_output currently
makes a distinction between frule arguments input vs. inputs and
output vs. outputs. I'm not sure if this is a good idea because it's confusing
but I did not change it to avoid breaking backwards compatibility.