Added t4161 - change case of an output file.
In Windows we do case-insensitive matching, so if we change the case of an output file, tup was matching the original name, and not updating correctly. For command strings and output filenames, we really want to do case sensitive matching so we can re-run the command and create the newly named output files as expected.

Fixes #159