Added t6072 - similar to t6071, but just remove the directory.
If a command writes to a file in another directory, then we remove the directory containing the command, we go through the tup_del_id_type() logic, which forcefully removes the directory and all of its nodes. However, since the generated file isn't contained in this directory, it ends up sitting around and holds onto the ghost directory via its srcid. Instead of trying to shoehorn generated file-removal into tup_del_id_type, we can instead "repase" the ghost directory and go through the same logic as if the Tupfile has just been removed (since it has). This way the generated files get removed as well, and our ghost directory can disappear.