Use unshare to create a new file-system namespace.
This allows us to chroot and/or mount the tup filesystem on top of the
real filesystem for the subprocesses, so we don't have the .tup/mnt path
showing up in the cwd for subprocesses. This means things like gcc
debugging & gcov should work out of the box. Additionally, we don't miss
dependencies if they use a full path instead of a relative path.

A user namespace is used to accomplish this without suid if possible,
but a suid tup is still supported for those kernels which don't have
user namespaces enabled.