!tup_preserve macro for including source files into variant builds
When using in-tree builds, the resulting build outputs may rely on
run-time files placed in the source tree. When using variants, it
may be helpful if such files are included in the variant build dir
as well. This patch provides a new macro "!tup_preserve", which can
be used to achieve this goal:

:foreach *.png |> !tup_preserve |>

Previously discussed here: https://goo.gl/zWsDEo

The patch also fixes the use of !tup_ln with variants.