Windows: Correctly read empty events.
If we got an empty event in the depfile, we'd have both event.len and
event.len2 == 0. We'd see that and skip to the next event. However, when
the strings are written out, we get a nul-byte for each. So even though
the lengths are 0, we still have to read the two nul-bytes. Failure to
do so results in errors like:

tup error: Size of 4915200 bytes is longer than the max filesize

This commit fixes at least that part of the error. However, the reason
for getting the 0-length events is because of the path canonicalization
routines in dllinject are failing from hitting the stupidly low
PATH_MAX.