dllinject: Hook NtQueryDirectoryFile, bump minimum version to Vista
It seems we need NtQueryDirectoryFile to hook certain file accesses with
Windows 10, otherwise tests like t4113 fail to get a dependency on a
ghost directory, and then fail to update properly. This requres the
GetFinalPathNameByHandle function, which means we have to specify a
minimum required Windows version. Since XP is no longer supported by
Microsoft, that should be fine. Additionally, we can remove the wow64.h
compatibility header since the definitions it provides are in the MinGW
headers given our minimum version.