Added t2197 - test for a Tupdefault ghost.
This re-works how we open Tupfiles, so that we only try to actually call
openat() if we have a tup_entry for the file. This avoids a lot of the
overhead of parsing tons of non-tup directories, since we don't have to
go to FUSE to check for Tupfile/Tupdefault/Tupfile.lua/Tupdefault.lua in
each one.