Database version 17 - split display & flags out of command strings.
Some commands are difficult or impossible to represent, since we store
the whole "^flags display^ command" directly in the database. For
example, it becomes difficult to copy a file with a carat in the name
(t2203), since the updater becomes confused where the separation lies
between the display and command. By making these separate columns in the
database, the parser can store each field directly, since it has more
context as to which is which.