Added t4184 - test for some dependency error messages.
If a command failed, tup could erroneously report that it failed to
process input dependencies, even if there were no input dependency
errors. The problem is that when the command fails, it generally doesn't
write the output file, and tup would report that as an input dependency
error. This should address the various cases of command / input / output
failures to report more sane messages.