Fix argument ordering ('tup -k foo' and 'tup foo -k' should both work)
Tup was originally implemented with 'tup cmd [cmdargs...]', where 'cmd'
would be something like 'init', 'upd', etc. Later, support for 'tup foo'
was added, where foo could be a filename, and was equivalent to 'tup upd
foo'. In this later form, it seems odd to force the flags to be after
the filename, so both 'tup -k foo' and 'tup foo -k' should set the
keep-going flag.

Everywhere that processes arguments now should do x through argc-1,
since the initial "tup" is removed, as well as the cmd argument if
available.