Use a single waiter thread.
The master fork loop was spawning a separate thread for each subprocess
to do a waitpid(). Instead we can put the process -> sid mapping
(child_wait_info) in a tree and use a single thread to do the waits. This
way we can join the single thread during shutdown to avoid errors with
helgrind caused by detached threads.