Generate version file using cmd.exe on Windows
On Windows, Tup doesn't execute commands through an Unix-style shell.
The rule used to generate `version.c` uses syntax that's specific to
Unix-style shells and thus fails on Windows. This commit provides an
alternative rule which is used when the host system on which Tup is
being built is Windows, and introduces a helper batch script which
contains the actual commands. The reason for the script is that cmd.exe
doesn't lend itself to one-liners, preferring instead to piss us off.