Added t6074 - generated directory ends up in create graph.
The generated directory here is a child of a normal directory. If that
normal directory is removed, it ends up as a ghost. When tup re-parses
the Tupfile, it turns the node into a generated directory, which is
still in the create list. Tup wasn't ignoring these, and instead
complained about unknown nodes. We should just skip any generated
directories in the create graph.