configs: Add GCC development LTO config template
Like gcc-dev.config, this template should be useful to those wanting to
develop tup. The difference in this one is the use of LTO, which can
help uncover errors in the code. At least here, tup did not build with
these settings; presumably, there's something to fix, and this template
is a step in that direction.