fix tup on musl
This patch fixes tup on linux systems based on musl libc.

In the master fork server, handle EINTR on read(2) so that the
delivery of SIGCHLD does not cause the parent to exit spuriously.

In the fuse filesystem, stop using AT_SYMLINK_NOFOLLOW in faccessat(2).
(Glibc transparently handles this case by calling fstatat(2) instead [1])
Fortunately, symlinks on Linux are always 0777, so we can do away
with using AT_SYMLINK_NOFOLLOW entirely.

[1]: https://man7.org/linux/man-pages/man2/faccessat.2.html