Remove open_notify() in favor of explicit parser dependencies.
This was a hack to make Windows behave more like FUSE where we can
track which files are opened by the parser to get Tupfile dependencies.
However, since we have complete control over the parser code, we can
just be explicit about adding a dependency when we open or access a
file. This makes it much easier on the server so that we don't have to
special-case stuff for the parser.

In the meantime we still need the parser_server for FUSE to support
run-scripts, but if those are removed in favor of a Python parser in the
future, we can remove all parser_server logic.