Update t2033 - allow for ghost dependencies on the dir or the file
With FUSE, if we do 'cat secret/ghost', and 'secret' does not exist as a
directory yet, then we only see the file access on 'secret', and not
'ghost'. However, with LD_PRELOAD we can see the access on the
underlying file. Both are valid ways of tracking the ghost, although
depending directly on secret/ghost is slightly more efficient. We should
be able to test for both cases though.