Force create graph mtime to -1
The ldpreload server may have a file with an mtime that becomes a ghost
directory with an intact mtime (instead of -1). This can throw off the
progress bar since it thinks it should use the mtime to calculate
progress instead of simply counting the number of nodes to parse. This
is hard to write a test case for, but was noticed in t2033.