Fixup t6049 - specify that overwritten files may need to be updated manually.
If a sub-process overwrites a file, we need to make different decisions
based on whether it was a ghost (we can remove it), a generated file
(reschedule the command), or a normal file (warn the user that they need
to take action). The FUSE server doesn't need to make these distinctions
because all files are written to a temporary directory first, but this
will help the shared library shim servers.