Fixup bootstrapping with ldpreload
Also I meant to mention in the last commit: The regular bootstrap.sh will build
with FUSE, while bootstrap-ldpreload.sh will build with LD_PRELOAD. Note that
this only applies for the bootstrapped version of tup in the build/ directory.
This tup executable is then used to build the tup tree using the Tupfiles, and
the server used by the final version of tup is driven by CONFIG_TUP_SERVER,
which defaults to FUSE. Setting CONFIG_TUP_SERVER=ldpreload and using
bootstrap-ldpreload.sh will avoid FUSE throughout the entire process.

The bootstrap-nofuse.sh script is separate, which uses the 'tup
generate' feature to create a temporary shell script so that libfuse is
only linked to, but not executed.