ldpreload: Remove syscall() wrapper.
This was only added because of a patch on a specific version of
coreutils in Fedora that implements renameat2() as a direct syscall. We
can just skip that test on Fedora instead.

The problem with this implementation is __NR_renameat2 isn't available
on all systems, and syscall() interacts negatively with jemalloc to
cause a hang.