Update t6044 - add check for missing mapping.
If we don't get an ACCESS_WRITE in the ldpreload server, we won't have a
mapping for a file, and without the mapping, we don't set the mtime.
This can cause tup to always think a file has been overwritten outside
of tup since the mtime never matches the expected value.