Added t4190, t4191 - realpath() tests.
This verifies that realpath() will give us a dependency on the file that
we're canonicalizing. This is made slightly tricky by the fact that when
we wrap realpath(), we have to get a specific version of the symbol that
handles a NULL resolved_path properly. This behavior is required by
certain programs, such as rustc.