Added t6076 - broken update with a circular dependency error.
This requires a fairly rare set of circumstances to hit. A rule has to
be parsed that creates an output file, then that output file must be
created outside of tup, so tup thinks the file will be resurrected.
Finally, the rule in the Tupfile must be updated so that the file is
listed as both an input and an output in the rule. Tup will skip a check
that the input is valid, and adds both an input and output link to the
same file, resulting in a circular dependency.