Added t6077, t6078 - broken update tests involving generated directories
This one took a while to track down, since a generated directory needs
to be made into a ghost, before being re-made into a generated directory
but not unflagged in the modify list, and then finally removed entirely
for it to fail. The problem is that the recursive generated directory
deletion and recursive ghost deletion wasn't calling delete_name_file(),
which actually handles unflagging the nodes. For the ghost case (t6078),
some flags are removed, but it is dependent on node type, with generated
nodes leaving the modify bit set. It went unnoticed for a while since
generated directories can't be flagged in create, and are only flagged
in modify in this specific case of becoming a ghost temporarily and
converting back.