Added t6079 - normal files need to be processed during partial updates.
The partial update works by building the PDAG and then pruning away all
nodes that don't point to the desired outputs. However, before we run a
command, we don't necessarily know all the inputs it will use, so it is
possible that we pruned out a regular file that is in the modify list.
In this case tup still thinks the file is modified, so on a future
update, tup will we-run the command that we just ran during the partial
update even though nothing has changed. Instead we need to make sure
that normal files are always processed even during a partial update. Any
downstream commands that were pruned are still flagged to run on a
future update.