Prune normal files from 'tup graph'
Since t6079 we don't prune normal files in prune_graph(), but when using
'tup graph' those files just add extra noise if we're using --prune. So
we need those nodes pruned from 'tup graph' but not from the updater so
that we can have correct mtimes but also easier-to-read graphs.