Added t4192 - test for --quiet with a command that fails due to dependency errors.
If we're displaying error messages, we always need to display the
banner, even in quiet mode. However, the check to display the banner
only considered the command's output, not the output coming from tup
itself for dependency errors and the like. We need to consider both for
the output to be readable.