Use link.sh to generate version every time tup is linked.
The tup-version.o file used to be built as a side-effect during linking
to ensure that we always had a fresh version from 'git describe', but
this was changed to support local Windows building to be a separate
step. Since we ignore files under .git during dependency detection, this
effectively meant that we'd generate a version once, and then never
update it unless version.o was manually removed. This caused a problem
on the CI system in particular, since it always builds incrementally,
and uses 'tup version' to determine the output filename. As a
consequence, Windows zips were incorrectly labeled for some time.

Instead we can use a helper script to generate the version as a
side-effect of linking simliar to before, so that we have correct
versions during incremental updates, and Windows local builds still
work.