Added t4193 - test for remove() function.
Usually files are deleted via unlink(), which we already track in the
LD_PRELOAD shim. However, some programs may call remove() instead, which
ends up calling the unlink syscall under the hood. We should have a
wrapper for remove() as well to detect these cases.

In particular, it seems that some versions of clang use remove() when
compiling assembly code, but not when compiling C/C++ code.