Added t4194 - stat() checks.
We didn't have any explicit stat() checks yet, so this test adds those.
This arose from using the clang compiler with the LD_PRELOAD checker,
which was behaving differently from the FUSE checker when calling stat()
on a file that already existed, but not subsequently opening a file. It
seems clang is just checking whether or not the path is a directory, and
only trying to open header files there if it is.

Note that clang behaves differently from gcc in this case. If we have a
commandline with '-Ifoo -Ibar' and #include "sub/header.h", then gcc will
do:

open("foo/sub/header.h") - ENOENT
open("bar/sub/header.h") - 0

While clang does:

stat("foo/sub")
 - if a directory, open("foo/sub/header.h") - ENOENT
stat("bar/sub")
 - if a directory, open("bar/sub/header.h") - 0

Clang's behavior presents a problem if "foo/sub" in this case is also
the name of say the final executable, because the stat appears to tup as
if the compilation command is trying to read from a future linker
output. The only real solution in this case is to name the directory
something different from the executable, or not pass in a -I argument
for the directory where the executable is created.