Perform less work when shutting down due to a signal
When we catch a signal, we kill all subprocesses and then process any
file access info that we captured while it was running. This is to
ensure that if the subprocess wrote a file that it wasn't supposed to
before getting a signal, we can still catch that and remove it if
necessary.

However, we don't need to process any file reads since all commands will
be rescheduled on the next update anyway, so handling those just wastes
time before tup returns to the commandline. Additionally, we can avoid
producing useless messages about files not being written when they were
expected to, since we don't expect a subprocess to write all files when
it is killed by a signal.