Added t6080 - generated -> normal conversion should remove links.
If a generated file is converted to a normal file (by touching the file
and also removing the rule that creates it), then we should no longer
be a member of any groups that the old generated file used to belong to.
We can do this by adding all outputs to the modify list and removing the
links before setting the type to a normal file.