Added t4186 - don't continually reparse if generated dirs are missing.
If a generated directory is missing, we don't have to reparse the
Tupfiles that write files there. Instead, we can just make sure that the
commands that would write those files are in the modify list and skip
parsing.

If we force-remove a generated directory, then we need to re-parse the
source Tupfiles in case they now need to throw an error about being
unable to write to that directory (since it may now be a normal file,
for example).

In contrast, if a normal directory is removed, we always need to
re-parse the source Tupfiles that write to it, since we either need to
convert it to a generated directory or throw an error.