Added t6081, t6082 - broken update converting a file with ghosts to a directory.
I think this went unnoticed for so long because only the ldpreload
server adds ghosts onto existing files, and this only occurs if such a
file is then removed and converted into something else. In this case,
since tup_del_id_force() wasn't actually removing the node in order
to preserve ghosts, we got an SQL reset error when trying to insert
the node again. We should just call back into the create function so the
new ghost can be handled properly if it exists.