Added t6083, t4203 - pthreads + forking tests
When building mozilla-central with tup, there was a small chance that a
particular rust build-script would cause tup to hang forever waiting for
a mutex in the ldplreoad shim. The process in question spawns multiple
threads, and the hang would occur when one thread would do some sort of
file access while another thread calls fork() to run a subprocess. The
first thread might end up in handle_file() inside the ldpreload shim and
take out the mutex, and then the child process of the fork() would think
that the mutex is locked, so when the execvp() gets called, it waits
forever trying to grab the mutex.

All we really need to do in this case is use pthread_atfork(), which was
intended to fix this situation. Apparently that may have some issues on
certain systems because the child handler isn't supposed to be allowed
to unlock a mutex that it is not technically the owner of, but it
appears to work on at least a few Linux machines that I've tested.