Added t2210 - fix for .gitignore files marked TUP_NODE_NORMAL
This is possibly the cause of issues #312, #313 and #328. If a user
creates a .gitignore file, then tup will create an entry in the database
with a type==TUP_NODE_NORMAL. This prevents tup from writing generated
files to the .gitignore file if later we add a .gitignore directive to
try to generated the file automatically.

Instead we need to convert the type from TUP_NODE_NORMAL to
TUP_NODE_GENERATED when the directive is added, and similarly convert it
back to TUP_NODE_NORMAL if there is still user-generated content
remaining when the directive is removed (if no content remains, we can
delete the file).