Ignore icecream's lock file.
If icecream (distributed C/C++ compilation) falls back to its local
build mode, it uses a lock file in /tmp/.icecream-* to run everything
one process at a time for some reason. Since the lock file is created
and never removed, tup flags it as a dependency error. We can just skip
this file using the same mechanism as ccache.