Reorganize struct path_list to use dir, dirlen instead of path.
The path element was a bit odd, since it was the directory part of the
path, and we had to overwrite mem in order to generate it. Instead we
can just use dir, dirlen to point into mem as appropriate. This should
make it easier to support exclusions explicitly in the path_list.