Move tup_db_get_outputs() query so we only query outputs once.
Both tup_db_check_actual_inputs() and tup_db_check_actual_outputs()
require the output tree, so if we move the outputs query to
write_files(), we can use a single db query for both. Since the
output_root is destroyed by compare_list_tree(), we still have to make a
copy of the tree, but we at least avoid the redundant query.